metrics/requirements.txt
certifi
numpy
opencv-python
pandas
Pillow
python-dateutil
pytz
PyYAML
six
torch
torchvision
tqdm
scikit-image
IPython
